Initial login roundtrip, and folder restructure.
Now making individual sites is much easier, as each specific site is
contained with all of its assets in the sites/ folder.  Also,
node_modules is added back for easy of deployment, and finally, the
initial parts of hashing a user's creds and making a roundtrip to the
server are complete.