Added backbone, refactored timeline & deps.
Added backbone and appropriate prereqs, sharing the libraries between
client and server.  Also refactored the timeline code to provide ids for
the elements it generates, and added UUIDs to all pieces of content,
which bind them to the Backbone models.

Experimented with Underscore alternatives, but turned them down due to
friction integrating them with RequireJS in the browser.  Too much
effort.