Now templates render multiple categories.
When the route is accessed, the categories desired can be specified in
the `info.categories` array.  Then, the appropriate categories will
be loaded, and accessible by the template being rendered.  This also
allowed for grabbing specific pieces of content via ajax, which sets the
`info.ajax` flag to `true` when a query string request is made from a
client.